Remove extra constructors for WebHookImpl
