+++ tcwebhooks-core/src/main/java/webhook/teamcity/WebHookFactory.java
@@ -5,6 +5,5 @@ import webhook.WebHookProxyConfig;
 import webhook.teamcity.settings.WebHookConfig;
 
 public interface WebHookFactory {
-	public abstract WebHook getWebHook();
 	public abstract WebHook getWebHook(WebHookConfig webhookConfig, WebHookProxyConfig pc);
 }
